Add screen resolution fields
I added the screen resolution fields to the `Engine` struct. This meant
I had to break up the big field into more smaller fields. Calculating
the start and end of the fields is error prone and tedious, and it might
be possible to make it simpler with some macro magic. I decided to just
add more tests to check if the offsets and lenghts make sense.

I plan to add a command line parameter for specifying the resolution.
That will make this tool much more useful since you will be able to
specify resolutions that are not supported by the game.