Add remaining fields
I added the remaning field that could be found by changing the settings
in the game. There are some values that cannot be changed from the
settings. I suspect the developers used the other settings for
debugging.

I noticed the gamma used two bytes but it does not align with a u16. I
must have done something wrong but I do not know what.

I used the `memoffset` crate to simplify the code and the unit tests. It
lets you read the offset of each field in a struct. I used to it verify
that each field was at the expected offset.

I have not added command line parameters for the remaining settings.
Those can be added later.